requirements.txt
aequitas==0.42.0
altair==4.1.0
altair-data-server==0.4.1
altair-saver==0.5.0
altair-viewer==0.4.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.3.0
bleach==5.0.0
blinker==1.5
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
commonmark==0.9.1
cryptography==37.0.4
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
dominate==2.7.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
Fiona==1.8.21
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
fontawesomefree==6.1.1
fonttools==4.33.3
geographiclib==1.52
geopandas==0.11.1
geopy==2.2.0
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.2
h11==0.13.0
html5lib==1.1
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.13.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.8
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.3.3
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyter-server==1.17.0
jupyterlab==3.4.2
jupyterlab-pygments==0.2.2
jupyterlab-server==2.14.0
jupyterlab-widgets==1.1.0
kaleido==0.2.1
kiwisolver==1.4.3
lightgbm==3.3.2
markdown2==2.3.5
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
millify==0.1.1
mistune==0.8.4
mizani==0.7.4
munch==2.5.0
mypy-extensions==0.4.3
nbclassic==0.3.7
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
notebook-shim==0.1.0
numpy==1.22.4
ohio==0.5.0
outcome==1.2.0
packaging==21.3
palettable==3.3.0
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
path==16.4.0
pathspec==0.9.0
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.5.2
plotly==5.8.0
plotly-geo==1.0.0
plotnine==0.9.0
portpicker==1.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pycodestyle==2.8.0
pycparser==2.21
pydeck==0.7.1
Pygments==2.12.0
Pympler==1.0.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyPDF2==2.10.0
pyproj==3.3.1
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywaffle==1.1.0
PyYAML==6.0
pyzmq==23.1.0
qtconsole==5.3.1
QtPy==2.1.0
reportlab==3.6.11
requests==2.27.1
rich==12.5.1
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
selenium==4.4.0
semver==2.13.0
Send2Trash==1.8.0
Shapely==1.8.2
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.37
stack-data==0.2.0
statsmodels==0.13.2
streamlit==1.11.1
tabulate==0.8.2
tenacity==8.0.1
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
tornado==6.1
traitlets==5.2.2.post1
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
validators==0.20.0
vega-datasets==0.9.0
visitor==0.1.3
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.2.2
widgetsnbextension==3.6.0
wsproto==1.1.0
xgboost==1.6.1
xhtml2pdf==0.2.2
zipp==3.8.1
